[Design] #40 - DropdownListView를 testView에 재생성중
