+++ PennyPack/ViewModel/ListModel.swift
@@ -18,15 +18,16 @@ class ListViewModel: ObservableObject {
     @Published var shoppingList: [ShoppingList] = []
     
     init() {
-        if shoppingList.isEmpty {
+//        if shoppingList.isEmpty {
             self.shoppingList = [
-                    ShoppingList(title: "January", isChoise: false, isPurchase: false),
-                    ShoppingList(title: "February", isChoise: false, isPurchase: false),
-                    ShoppingList(title: "March", isChoise: false, isPurchase: false),
-                    ShoppingList(title: "April", isChoise: false, isPurchase: false),
-                    ShoppingList(title: "May", isChoise: false, isPurchase: false)
+                        ShoppingList(title: "January", isChoise: false, isPurchase: false),
+                        ShoppingList(title: "February", isChoise: false, isPurchase: false),
+                        ShoppingList(title: "March", isChoise: false, isPurchase: false),
+                        ShoppingList(title: "April", isChoise: false, isPurchase: false),
+                        ShoppingList(title: "May", isChoise: false, isPurchase: false)
             ]
-        }
+//        }
+        saveShoppingListToUserDefaults() // 테스트할때만 있는거고 나중에 앱 출시할땐 save 지워야함.
         loadShoppingListFromUserDefaults()
     }
